Clean up compiler warnings (primarily on Windows) (#720)
* Fix RelWithDebInfo and MinSizeRel targets on Windows

* Use custom CUDA_INFINITY instead of INFINITY

* Use __COUNTER__ in LOG_TIMER macros to avoid name collisions

* Suppress flood of benign CUDA/MSVC warnings on Windows

Silences NVCC #221 (INFINITY overflow) and MSVC C4267 (size_t conversion)
warnings from standard library headers in CUDA compilation.

* Explicit casts to suppress conversion warnings

* Discard [[nodiscard]] return values to suppress C4834

* Avoiding more compilation warnings

* Declaration hides previous variable

* Removed garbage variables and other minor warnings

* Fix log concatenation

* Fix -Wformat-security warning in CUDA logger

* #include <cfloat> to use FLT_MAX in tensor_ops

* <cfloat> include to fix FLT_MAX undefined (problem on Github runner only?)

* Replace (void) casts with proper error handling and logging

* Using static cast instead of classic C casting

* Clearer variable name from ft to invFlash

* Ride of more compilation warnings

+++ src/io/nvcodec_image_loader.cpp
@@ -825,7 +825,7 @@ namespace lfs::io {
             decoder,
             code_streams.data(),
             nv_images.data(),
-            batch_size,
+            static_cast<int>(batch_size),
             &decode_params,
             &decode_future);
 
@@ -990,7 +990,7 @@ namespace lfs::io {
             decoder,
             code_streams.data(),
             nv_images.data(),
-            batch_size,
+            static_cast<int>(batch_size),
             &decode_params,
             &decode_future);
